requirements.txt
# Core
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langgraph>=0.2.0
langchain-text-splitters>=0.3.0

# PDF Processing
pymupdf>=1.24.0          # fitz - fast text + image extraction
camelot-py[cv]>=0.11.0   # table extraction
unstructured[pdf]>=0.15.0
pdfplumber>=0.11.0

# Vector DB
chromadb>=0.5.0

# Embeddings & LLM
google-generativeai>=0.8.0

# UI
streamlit>=1.38.0
streamlit-chat>=0.1.1

# Utils
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
Pillow>=10.0.0
pytesseract>=0.3.10       # OCR for image tables
requests>=2.31.0
tqdm>=4.66.0